# backend/openapi.json
{
  "components": {
    "schemas": {
      "AcceptRequest": {
        "properties": {
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "AcceptRequest",
        "type": "object"
      },
...
      "InvitationView": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "expiresAt",
          "createdAt"
        ],
        "title": "InvitationView",
        "type": "object"
      },
      "IssuedInvitationView": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "id",
          "expiresAt",
          "createdAt",
          "token"
        ],
        "title": "IssuedInvitationView",
        "type": "object"
      },
      "MembershipView": {
        "properties": {
          "role": {
            "title": "Role",
            "type": "string"
          },
          "spaceId": {
            "format": "uuid",
            "title": "Spaceid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "spaceId",
          "role",
          "status"
        ],
        "title": "MembershipView",
        "type": "object"
      },
...
      "ProblemDetails": {
        "description": "Der Rumpf jeder Fehlerantwort.\n\nDasselbe Modell erzeugt die Antwort und beschreibt sie im OpenAPI-\nVertrag. Getrennt gepflegt wuerden beide irgendwann auseinanderlaufen,\nund der Vertrag beschriebe dann einen Fehler, den es so nicht gibt.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Application Core. OpenAPI ist der verbindliche Vertrag.",
    "title": "SideBySide Next",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/invitations/accept": {
      "post": {
        "description": "Eine Einladung annehmen.\n\nAusserhalb von /spaces/..., weil der Aufrufer den Space noch nicht\nkennt - der Token bestimmt ihn.",
        "operationId": "accept_invitation_api_v1_invitations_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentifizierung fehlt, ist ungueltig oder die Sitzung ist abgelaufen."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Die Anfrage kollidiert mit dem aktuellen Zustand der Ressource."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Anfrageparameter oder fachliche Eingaben sind ungueltig."
          }
        },
        "summary": "Accept Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
...
    "/api/v1/spaces/{spaceId}/invitations": {
      "get": {
        "description": "Die offenen Einladungen. Ohne Token - der ist einmalig gewesen.",
        "operationId": "list_invitations_api_v1_spaces__spaceId__invitations_get",
        "parameters": [
          {
            "in": "path",
            "name": "spaceId",
            "required": true,
            "schema": {
              "title": "Spaceid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvitationView"
                  },
                  "title": "Response List Invitations Api V1 Spaces  Spaceid  Invitations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentifizierung fehlt, ist ungueltig oder die Sitzung ist abgelaufen."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Die Ressource existiert nicht oder ist fuer den Aufrufer nicht sichtbar."
          }
        },
        "summary": "List Invitations",
        "tags": [
          "invitations"
        ]
      },
      "post": {
        "operationId": "create_invitation_api_v1_spaces__spaceId__invitations_post",
        "parameters": [
          {
            "in": "path",
            "name": "spaceId",
            "required": true,
            "schema": {
              "title": "Spaceid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedInvitationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentifizierung fehlt, ist ungueltig oder die Sitzung ist abgelaufen."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Die Ressource existiert nicht oder ist fuer den Aufrufer nicht sichtbar."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Die Anfrage kollidiert mit dem aktuellen Zustand der Ressource."
          }
        },
        "summary": "Create Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/v1/spaces/{spaceId}/invitations/{invitationId}": {
      "delete": {
        "operationId": "revoke_invitation_api_v1_spaces__spaceId__invitations__invitationId__delete",
        "parameters": [
          {
            "in": "path",
            "name": "invitationId",
            "required": true,
            "schema": {
              "title": "Invitationid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "spaceId",
            "required": true,
            "schema": {
              "title": "Spaceid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentifizierung fehlt, ist ungueltig oder die Sitzung ist abgelaufen."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Die Ressource existiert nicht oder ist fuer den Aufrufer nicht sichtbar."
          }
        },
        "summary": "Revoke Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
...
  }
}